Fix compiler type warnings.

+++ thread_pool.h
@@ -13,8 +13,8 @@ namespace pk
 {
 
 typedef uint32_t thread_pool_t;
-typedef uint32_t job_t;
-typedef uint32_t job_group_t;
+typedef uint64_t job_t;
+typedef uint64_t job_group_t;
 
 #define INVALID_THREAD_POOL ( thread_pool_t( -1 ) )
 #define INVALID_JOB ( job_t( -1 ) )
